+++ package-lgtseek/pipeline_builder/perl/create_lgt_pipeline_config.pl
@@ -344,8 +344,10 @@ sub main {
 
 		# If recipient is infected with LGT, change mpileup to use LGT-infected BAM list
 		if ($lgt_infected) {
-			$config{'sam2fasta fasta_r'}->{'$;INPUT_FILE$;'} ='$;REPOSITORY_ROOT$;/output_repository/samtools_merge/$;PIPELINEID$;_lgt_infected/samtools_merge.bam.list';
-			$config{'gather_lgtseek_files'}->{'RECIPIENT_LGT_BAM_OUTPUT'} = '$;REPOSITORY_ROOT$;/output_repository/filter_dups_lc_seqs/$;PIPELINEID$;_lgt_infected/filter_dups_lc_seqs.bam.list' if $included_subpipelines{'post'};
+			$config{'sam2fasta fasta_r'}->{'$;INPUT_FILE$;'} ='$;REPOSITORY_ROOT$;/output_repository/samtools_merge/$;PIPELINEID$;_lgt_infected_r/samtools_merge.bam.list';
+			$config{'sam2fasta fasta_d'}->{'$;INPUT_FILE$;'} ='$;REPOSITORY_ROOT$;/output_repository/samtools_merge/$;PIPELINEID$;_lgt_infected_d/samtools_merge.bam.list';
+			$config{'gather_lgtseek_files'}->{'RECIPIENT_LGT_BAM_OUTPUT'} = '$;REPOSITORY_ROOT$;/output_repository/samtools_merge/$;PIPELINEID$;_lgt_infected_r/samtools_merge.bam.list' if $included_subpipelines{'post'};
+			$config{'gather_lgtseek_files'}->{'DONOR_LGT_BAM_OUTPUT'} = '$;REPOSITORY_ROOT$;/output_repository/samtools_merge/$;PIPELINEID$;_lgt_infected_d/samtools_merge.bam.list' if $included_subpipelines{'post'};
 		} else {
             # For good donor, good LGT-free recipient, donor and recipient reads are the same so we just run split_multifasta once
             $config{'blastn_plus nt_r'}->{'$;INPUT_FILE_LIST$;'} = '$;REPOSITORY_ROOT$;/output_repository/split_multifasta/$;PIPELINEID$;_fasta_d/split_multifasta.fsa.list';