+++ package-lgtseek/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+1/18/17
+* Fixing bug in 'lgt_infected' use-case where only the recipient reads of lgt_infected cases were being output, and not the donor reads
+* Consequently this requires some of the pipeline to be rewritten
+
 11/29/17
 * Renamed 'lgt\_bwa' to 'bwa\_aln' since there may be a future case where we implement bwa-mem instead.